scraper-service/requirements.txt
selenium==4.15.2
beautifulsoup4==4.12.2
apscheduler==3.10.4
python-dotenv==1.0.0
requests==2.31.0
lxml==4.9.3
webdriver-manager==4.0.1
python-dateutil==2.8.2